Resolve comments, move function to utils file
